+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - AccountId support for ECDSA alias accounts
 - ContractId.to_evm_address() method for EVM compatibility
 - consumeLargeData() function in StatefulContract
+- example script for Token Airdrop
 
 ### Changed
 - Extract _build_proto_body() from build_transaction_body() in every transaction